pratica3/requirements.txt
pillow==11.2.1 
numpy==2.2.4